# src/main/resources/openapi/court-list-publisher.openapi.yml
openapi: 3.0.3
servers:
  - description: APIHub API Auto Mocking Scheduling and Listing court-schedule
    url: https://virtserver.swaggerhub.com/HMCTS-DTS//api-cp-crime-court-list-publisher/0.0.0
info:
  title: Common Platform API Crime Court List Publisher
  description: API for court list publishing service
  version: 0.0.0
  contact:
    email: no-reply@hmcts.com
  license:
    name: MIT
    url: 'https://opensource.org/licenses/MIT'
paths:
  /api/court-list-publish/publish:
    post:
      summary: Publish court list
      description: Creates or updates a court list publish status
      operationId: publishCourtList
      tags:
        - court-list-publish
      requestBody:
        required: true
        content:
          application/vnd.courtlistpublishing-service.publish.post+json:
            schema:
              $ref: "#/components/schemas/CourtListPublishRequest"
      responses:
        '200':
          description: Successfully created or updated court list publish status
          content:
            application/vnd.courtlistpublishing-service.publish.post+json:
              schema:
                $ref: "#/components/schemas/CourtListPublishResponse"
            application/json:
              schema:
                $ref: "#/components/schemas/CourtListPublishResponse"
        '400':
          description: Bad request - invalid request body or missing required fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/court-list-publish/publishCourtList:
    post:
      summary: Publish SJP court list to CaTH
      description: |
        Triggers publishing of an SJP (Single Justice Procedure) court list to Courts and Tribunals Hearings (CaTH).
        Equivalent to the flow in PubHub: consume SJP list type, transform to CaTH format, and POST to Azure APIM.
        Use listType to specify either the SJP Pending cases public list (SJP_PUBLISH_LIST) or the SJP Press transparency list (SJP_PRESS_LIST).
      operationId: publishSjpCourtList
      tags:
        - court-list-publish
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishCourtListRequest"
          application/vnd.courtlistpublishing-service.publishcourtlist.post+json:
            schema:
              $ref: "#/components/schemas/PublishCourtListRequest"
      responses:
        '200':
          description: SJP court list publish request accepted or completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishCourtListResponse"
            application/vnd.courtlistpublishing-service.publishcourtlist.post+json:
              schema:
                $ref: "#/components/schemas/PublishCourtListResponse"
        '400':
          description: Bad request - invalid or missing listType
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/court-list-publish/publish-status:
    get:
      summary: Get court list publish statuses
      description: |
        Retrieves court list publish statuses. Can be queried by:
        - courtListId (UUID), OR
        - courtCentreId (required) and publishDate (required) with optional courtListType
      operationId: findCourtListPublishStatus
      tags:
        - court-list-publish
      parameters:
        - in: query
          name: courtListId
          required: false
          description: The UUID of the court list. If provided, other parameters are ignored.
          schema:
            type: string
            format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        - in: query
          name: courtCentreId
          required: false
          description: The UUID of the court centre. Required if courtListId is not provided.
          schema:
            type: string
            format: uuid
          example: "123e4567-e89b-12d3-a456-426614174001"
        - in: query
          name: publishDate
          required: false
          description: The publish date in format 'YYYY-MM-DD'. Required if courtListId is not provided.
          schema:
            type: string
            format: date
          example: "2024-01-15"
        - in: query
          name: courtListType
          required: false
          description: The type of court list. Optional when querying by courtCentreId and publishDate.
          schema:
            $ref: "#/components/schemas/CourtListType"
          example: "STANDARD"
      responses:
        '200':
          description: Successfully retrieved court list publish statuses. Returns an empty array if no matching records are found.
          content:
            application/vnd.courtlistpublishing-service.publish.get+json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CourtListPublishResponse"
        '400':
          description: Bad request - invalid parameters. Either courtListId must be provided, or both courtCentreId and publishDate must be provided.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/court-list-publish/courtlistdata:
    get:
      summary: Get court list data
      description: |
        Returns court list data as JSON from listing and reference data.
        Listing payload with optional ouCode/courtId from reference data.
      operationId: getCourtlistData
      tags:
        - court-list-publish
      parameters:
        - in: query
          name: courtCentreId
          required: true
          description: Reference data ID for the court centre
          schema:
            type: string
            format: uuid
          example: "f8254db1-1683-483e-afb3-b87fde5a0a26"
        - in: query
          name: courtRoomId
          required: false
          description: Reference data ID for the court room (optional)
          schema:
            type: string
            format: uuid
        - in: query
          name: listId
          required: true
          description: Court list type (e.g. PUBLIC, STANDARD, USHERS_CROWN, USHERS_MAGISTRATE). PRISON not supported.
          schema:
            $ref: "#/components/schemas/CourtListType"
          example: "STANDARD"
        - in: query
          name: startDate
          required: true
          description: Hearing start date in format 'YYYY-MM-DD'
          schema:
            type: string
            format: date
          example: "2024-01-15"
        - in: query
          name: endDate
          required: true
          description: Hearing end date in format 'YYYY-MM-DD'
          schema:
            type: string
            format: date
          example: "2024-01-15"
        - in: query
          name: restricted
          required: false
          description: For standard list, set true to get restricted list. Default false.
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Court list data (listing payload with optional ouCode/courtId from reference data)
          content:
            application/vnd.courtlistpublishing-service.publish.courtlistdata+json:
              schema:
                $ref: "#/components/schemas/CourtListData"
        '403':
          description: Forbidden (e.g. PRISON listId not supported)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '400':
          description: Bad request - missing or invalid parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
    SjpListType:
      description: SJP court list type for publishing to CaTH (Single Justice Procedure)
      type: string
      enum:
        - SJP_PUBLISH_LIST
        - SJP_PRESS_LIST
      example: "SJP_PUBLISH_LIST"
    PublishCourtListRequest:
      description: Request to publish an SJP court list to CaTH. listType determines which list is published (public/pending list or press transparency list).
      type: object
      required:
        - listType
      properties:
        listType:
          $ref: "#/components/schemas/SjpListType"
    PublishCourtListResponse:
      description: Response after accepting or completing an SJP court list publish request
      type: object
      properties:
        status:
          type: string
          description: Status of the publish request
          example: "ACCEPTED"
        listType:
          $ref: "#/components/schemas/SjpListType"
        message:
          type: string
          description: Optional human-readable message
          example: "SJP court list publish request accepted"
    CourtListType:
      description: The type of court list
      type: string
      enum:
        - ALPHABETICAL
        - PUBLIC
        - STANDARD
        - BENCH
        - JUDGE
        - USHERS_CROWN
        - USHERS_MAGISTRATE
        - ONLINE_PUBLIC
        - PRISON
        - WARN
        - DRAFT
        - FINAL
        - FIRM
      example: "STANDARD"
    Status:
      description: A generic status that can be used for various operations
      type: string
      enum:
        - REQUESTED
        - SUCCESSFUL
        - FAILED
      example: "REQUESTED"
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Machine-readable error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          additionalProperties: true
          description: Additional error context
        timestamp:
          type: string
          format: date-time
        traceId:
          type: string
          description: Unique identifier for error tracing
...
    CourtListPublishRequest:
      description: Request to publish a court list. The courtListId is generated internally by the backend, and publishStatus is set to COURT_LIST_REQUESTED.
      type: object
      required:
        - courtCentreId
        - courtListType
        - startDate
        - endDate
      properties:
        courtCentreId:
          description: The UUID of the court centre
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174001"
        startDate:
          description: Hearing start date in format 'YYYY-MM-DD'
          type: string
          format: date
          example: "2024-01-15"
        endDate:
          description: Hearing end date in format 'YYYY-MM-DD'
          type: string
          format: date
          example: "2024-01-15"
        courtListType:
          $ref: "#/components/schemas/CourtListType"
    CourtListPublishResponse:
      description: Response containing court list publish status information
      type: object
      required:
        - courtListId
        - courtCentreId
        - publishStatus
        - courtListType
      properties:
        courtListId:
          description: The UUID of the court list
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        courtCentreId:
          description: The UUID of the court centre
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174001"
        publishStatus:
          $ref: "#/components/schemas/Status"
        fileStatus:
          $ref: "#/components/schemas/Status"
        courtListType:
          $ref: "#/components/schemas/CourtListType"
        lastUpdated:
          description: The timestamp when the publish status was last updated
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        fileUrl:
          description: The name of the court list file or its URL
          type: string
          example: "https://sastecourtlistpublisher.blob.core.windows.net/courtpublisher-blob-container/Online%20Public%20court%20list%20-%20Lavender%20Hill%20Magistrates'%20Court%2C%20All%20courtrooms%20-%2026-01-2026.pdf?st=2026-01-26T13:22:49Z&se=2026-12-31T21:37:49Z&moreParam=111"
        fileId:
          description: The UUID of the court list file in blob storage or document store
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174002"
        fileErrorMessage:
          description: Error message if file upload failed
          type: string
          example: "Failed to upload file"
        publishErrorMessage:
          description: Error message if publishing failed
          type: string
          example: "Failed to publish file"
        publishDate:
          description: The date when the court list was published
          type: string
          format: date
          example: "2024-01-15"
    CourtListData:
      description: Court list data from listing and reference data. Listing payload plus optional ouCode/courtId from reference data.
      type: object
      properties:
        listType:
          type: string
          example: "public"
        courtCentreName:
          type: string
        courtCentreDefaultStartTime:
          type: string
          description: Default start time for the court centre (e.g. "10:00:00")
          example: "10:00:00"
        courtCentreAddress1:
          type: string
        courtCentreAddress2:
          type: string
        welshCourtCentreName:
          type: string
        welshCourtCentreAddress1:
          type: string
        welshCourtCentreAddress2:
          type: string
        hearingDates:
          type: array
          items:
            $ref: "#/components/schemas/HearingDateEntry"
        templateName:
          type: string
          example: "PublicCourtList"
        ouCode:
          type: string
          description: Organisation unit code from reference data
        oucodeL1Code:
          type: string
          description: Organisation unit L1 code from reference data
        courtId:
          type: string
          description: Court centre ID from reference data
        courtIdNumeric:
          type: string
          description: Court ID numeric from reference data
    HearingDateEntry:
      description: Court list data for a single hearing date
      type: object
      properties:
        hearingDate:
          type: string
          format: date
          description: The hearing date (YYYY-MM-DD)
        hearingDateWelsh:
          type: string
          description: Welsh translation of the hearing date
        courtRooms:
          type: array
          items:
            $ref: "#/components/schemas/CourtRoomEntry"
    CourtRoomEntry:
      description: Court room with timeslots and hearings
      type: object
      properties:
        courtRoomName:
          type: string
        welshCourtRoomName:
          type: string
        judiciaryNames:
          type: string
        welshJudiciaryNames:
          type: string
        timeslots:
          type: array
          items:
            $ref: "#/components/schemas/Timeslot"
    Timeslot:
      description: A time slot containing one or more hearings
      type: object
      properties:
        hearings:
          type: array
          items:
            $ref: "#/components/schemas/Hearing"
    Hearing:
      description: A single hearing within the court list
      type: object
      properties:
        id:
          type: string
          format: uuid
        sequence:
          type: integer
        reportingRestrictionReason:
          type: string
        welshReportingRestrictionReason:
          type: string
        startTime:
          type: string
          description: Time in HH:mm format
        hearingType:
          type: string
        welshHearingType:
          type: string
        caseNumber:
          type: string
        caseId:
          type: string
          format: uuid
        prosecutorType:
          type: string
        hearingPublicListNote:
          type: string
        defendants:
          type: array
          items:
            $ref: "#/components/schemas/Defendant"
        adjournedHearingDate:
          type: string
    Defendant:
      description: Defendant in a hearing
      type: object
      properties:
        id:
          type: string
          format: uuid
        organisationName:
          type: string
        welshOrganisationName:
          type: string
        firstName:
          type: string
        surname:
          type: string
        welshSurname:
          type: string
        dateOfBirth:
          type: string
        age:
          type: string
        nationality:
          type: string
        address:
          $ref: "#/components/schemas/CourtListAddress"
        reportingRestrictions:
          type: array
          items:
            type: object
          description: List of reporting restriction entries
        offences:
          type: array
          items:
            $ref: "#/components/schemas/Offence"
    CourtListAddress:
      description: Address used in court list data (defendant address etc.)
      type: object
      properties:
        address1:
          type: string
        address2:
          type: string
        address3:
          type: string
        address4:
          type: string
        address5:
          type: string
        postcode:
          type: string
    Offence:
      description: Offence linked to a defendant
      type: object
      properties:
        id:
          type: string
          format: uuid
        offenceTitle:
          type: string
        welshOffenceTitle:
          type: string
        offenceWording:
          type: string